feat: 검색 기능 close #3

+++ Memo_programmers/Model/CoreData/CoreDataManager.swift
@@ -7,28 +7,24 @@
 //
 import Foundation
 import CoreData
-
-import Foundation
-import CoreData
 import UIKit
 
 class CoreDataManager {
-  
+
   static let sharedManager = CoreDataManager()
-  private init() {} 
-  
+  private init() {}
+
   lazy var persistentContainer: NSPersistentContainer = {
-    
+
     let container = NSPersistentContainer(name: "Memo_programmers")
-    container.loadPersistentStores(completionHandler: { (storeDescription, error) in
-      
+    container.loadPersistentStores(completionHandler: { (_, error) in
+
       if let error = error as NSError? {
         fatalError("Unresolved error \(error), \(error.userInfo)")
       }
     })
     return container
   }()
-  
 
   func saveContext () {
     let context = CoreDataManager.sharedManager.persistentContainer.viewContext
@@ -41,27 +37,28 @@ class CoreDataManager {
       }
     }
   }
-  
+
   // Insert
-  func add(newMemo: MemoData) -> (Bool, Error?){
+  func add(newMemo: MemoData) -> (Bool, Error?) {
     let managedContext = CoreDataManager.sharedManager.persistentContainer.viewContext
-    
-    
+
     var coredataTypeArr: [Images] = []
-    for i in newMemo.imageArray ?? [] {
-      let image = (NSEntityDescription.insertNewObject(forEntityName: "Images", into: managedContext) as! Images)
-      image.identifier = i.image.jpegData(compressionQuality: 0.8)!
-      image.date = i.date
-      coredataTypeArr.append(image)
+    for image in newMemo.imageArray ?? [] {
+      if let imageData = (NSEntityDescription.insertNewObject(forEntityName: "Images",
+                                                              into: managedContext) as? Images) {
+        imageData.identifier = image.image.jpegData(compressionQuality: 0.8)!
+        imageData.date = image.date
+        coredataTypeArr.append(imageData)
+      }
     }
-    
+
     let memo = Memo(context: managedContext)
     memo.title = newMemo.title
     memo.memo = newMemo.memo
     memo.date = newMemo.date
     memo.identifier = newMemo.identifier
     memo.images = NSSet(array: coredataTypeArr)
-    
+
     do {
       try managedContext.save()
       return (true, nil)
@@ -70,41 +67,42 @@ class CoreDataManager {
       return (false, error)
     }
   }
-  
+
   // delete
   func delete(identifier: UUID) -> (Bool, Error?) {
 
     let managedContext = CoreDataManager.sharedManager.persistentContainer.viewContext
-    
+
     let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: "Memo")
     fetchRequest.returnsObjectsAsFaults = false
     let predicate = NSPredicate(format: "identifier == %@", identifier as CVarArg)
     fetchRequest.predicate = predicate
 
     do {
-      
+
       let fetchResults: Array = try managedContext.fetch(fetchRequest)
-      
+
       for fetchResult in fetchResults {
-        let managedObject = fetchResult as! NSManagedObject
-        
-        managedContext.delete(managedObject)
+        if let managedObject = fetchResult as? NSManagedObject {
+
+          managedContext.delete(managedObject)
+        }
       }
-      } catch let error as NSError {
+    } catch let error as NSError {
       print(error)
       return (false, error)
     }
-    
+
     do {
       try managedContext.save()
       return (true, nil)
-      } catch let error as NSError {
+    } catch let error as NSError {
       return (false, error)
     }
   }
-  
+
   // update
-  func update(updateMemo: MemoData) -> (Bool, Error?){
+  func update(updateMemo: MemoData) -> (Bool, Error?) {
     let managedContext = CoreDataManager.sharedManager.persistentContainer.viewContext
     let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: "Memo")
     fetchRequest.returnsObjectsAsFaults = false
@@ -112,33 +110,36 @@ class CoreDataManager {
     fetchRequest.predicate = predicate
     fetchRequest.fetchLimit = 1
     var coredataTypeArr: [Images] = []
-    for i in updateMemo.imageArray ?? [] {
-      let image = (NSEntityDescription.insertNewObject(forEntityName: "Images", into: managedContext) as! Images)
-      image.identifier = i.image.jpegData(compressionQuality: 0.8)!
-      image.date = i.date
-      coredataTypeArr.append(image)
+    for image in updateMemo.imageArray ?? [] {
+      if let imageData = (NSEntityDescription.insertNewObject(forEntityName: "Images",
+                                                              into: managedContext) as? Images) {
+        imageData.identifier = image.image.jpegData(compressionQuality: 0.8)!
+        imageData.date = image.date
+        coredataTypeArr.append(imageData)
+      }
     }
     do {
       let fetchResult = try managedContext.fetch(fetchRequest).first
-      let managedObject = fetchResult as! NSManagedObject
-      managedObject.setValue(updateMemo.title, forKey: "title")
-      managedObject.setValue(updateMemo.memo, forKey: "memo")
-      managedObject.setValue(updateMemo.date, forKey: "date")
-      managedObject.setValue(updateMemo.modifyDate, forKey: "modifyDate")
-      managedObject.setValue(NSSet(array: coredataTypeArr), forKey: "images")
+      if let managedObject = fetchResult as? NSManagedObject {
+        managedObject.setValue(updateMemo.title, forKey: "title")
+        managedObject.setValue(updateMemo.memo, forKey: "memo")
+        managedObject.setValue(updateMemo.date, forKey: "date")
+        managedObject.setValue(updateMemo.modifyDate, forKey: "modifyDate")
+        managedObject.setValue(NSSet(array: coredataTypeArr), forKey: "images")
+      }
       do {
         try managedContext.save()
         return (true, nil)
-        } catch let error as NSError {
+      } catch let error as NSError {
         return (false, error)
       }
-      } catch let error as NSError {
+    } catch let error as NSError {
       return (false, error)
     }
   }
 
   // fetch all memo
-  func fetchAllMemos() -> [Memo]?{
+  func fetchAllMemos() -> [Memo]? {
     let managedContext = CoreDataManager.sharedManager.persistentContainer.viewContext
     let fetchRequest2 = NSFetchRequest<Memo>(entityName: "Memo")
     // unit test check
@@ -148,7 +149,8 @@ class CoreDataManager {
       let orderType = userPreferences.getOrderType()
       switch orderType {
       case .title:
-        fetchRequest2.sortDescriptors = [NSSortDescriptor(key: "title", ascending: true, selector: #selector(NSString.localizedCaseInsensitiveCompare))]
+        fetchRequest2.sortDescriptors = [NSSortDescriptor(key: "title", ascending: true,
+                                                          selector: #selector(NSString.localizedCaseInsensitiveCompare))]
       case .createDate:
         fetchRequest2.sortDescriptors = [NSSortDescriptor(key: "date", ascending: false)]
       case .modifyDate:
@@ -162,19 +164,18 @@ class CoreDataManager {
       print("Could not fetch. \(error), \(error.userInfo)")
       return nil
     }
-    
+
   }
 
-  
   // flush
   func flushData() {
-    
-    let fetchRequest:NSFetchRequest<NSFetchRequestResult> = NSFetchRequest<NSFetchRequestResult>(entityName: "Memo")
-    let objs = try! CoreDataManager.sharedManager.persistentContainer.viewContext.fetch(fetchRequest)
-    for case let obj as NSManagedObject in objs {
-      CoreDataManager.sharedManager.persistentContainer.viewContext.delete(obj)
+
+    let fetchRequest: NSFetchRequest<NSFetchRequestResult> = NSFetchRequest<NSFetchRequestResult>(entityName: "Memo")
+    if let objs = try? CoreDataManager.sharedManager.persistentContainer.viewContext.fetch(fetchRequest) {
+      for case let obj as NSManagedObject in objs {
+        CoreDataManager.sharedManager.persistentContainer.viewContext.delete(obj)
+      }
+      try? CoreDataManager.sharedManager.persistentContainer.viewContext.save()
     }
-    
-    try! CoreDataManager.sharedManager.persistentContainer.viewContext.save()
   }
 }
